+++ src/Builder/AbstractRequestBuilder.php
@@ -32,6 +32,7 @@ use OnlinePayments\Sdk\Domain\RedirectionData;
 use OnlinePayments\Sdk\Domain\RedirectPaymentMethodSpecificInput;
 use OnlinePayments\Sdk\Domain\RedirectPaymentProduct5402SpecificInput;
 use OnlinePayments\Sdk\Domain\RedirectPaymentProduct5403SpecificInput;
+use OnlinePayments\Sdk\Domain\RedirectPaymentProduct3112SpecificInput;
 use OnlinePayments\Sdk\Domain\Shipping;
 use OnlinePayments\Sdk\Domain\SurchargeSpecificInput;
 use RandomLib\Factory;
@@ -54,6 +55,7 @@ abstract class AbstractRequestBuilder implements PaymentRequestBuilderInterface
     const PRODUCT_ID_MEALVOUCHER = 5402;
     const PRODUCT_ID_CVCO = 5403;
     const PRODUCT_ID_PLEDG = 5300;
+    const PRODUCT_ID_ILLICADO = 3112;
 
     const PHONE_NUMBER_MAX_CHARS = 15;
 
@@ -142,7 +144,8 @@ abstract class AbstractRequestBuilder implements PaymentRequestBuilderInterface
         }
         if ($this->idProduct == self::PRODUCT_ID_MEALVOUCHER
             || (int) $this->idProduct === self::PRODUCT_ID_CVCO
-            || (int) $this->idProduct === self::PRODUCT_ID_PLEDG) {
+            || (int) $this->idProduct === self::PRODUCT_ID_PLEDG
+            || (int) $this->idProduct === self::PRODUCT_ID_ILLICADO) {
             $redirectPaymentMethodSpecificInput->setRequiresApproval(false);
         } else {
             $redirectPaymentMethodSpecificInput->setRequiresApproval(
@@ -163,6 +166,10 @@ abstract class AbstractRequestBuilder implements PaymentRequestBuilderInterface
         $product5403SpecificInput = new RedirectPaymentProduct5403SpecificInput();
         $product5403SpecificInput->setCompleteRemainingPaymentAmount(true);
         $redirectPaymentMethodSpecificInput->setPaymentProduct5403SpecificInput($product5403SpecificInput);
+
+        $product3112SpecificInput = new RedirectPaymentProduct3112SpecificInput();
+        $product3112SpecificInput->setCompleteRemainingPaymentAmount(true);
+        $redirectPaymentMethodSpecificInput->setPaymentProduct3112SpecificInput($product3112SpecificInput);
 
         return $redirectPaymentMethodSpecificInput;
     }