+++ src/Presenter/TransactionPresenter.php
@@ -34,6 +34,7 @@ use WorldlineOP\PrestaShop\Utils\Tools;
  */
 class TransactionPresenter implements PresenterInterface
 {
+    const STATUS_PAYMENT_CREATED = 'CREATED';
     const STATUS_REFUND_REQUESTED = 'REFUND_REQUESTED';
     const STATUS_CAPTURE_REQUESTED = 'CAPTURE_REQUESTED';
     const STATUS_PAYMENT_CAPTURED = 'CAPTURED';
@@ -85,6 +86,7 @@ class TransactionPresenter implements PresenterInterface
         if ($paymentDetails) {
             foreach ($paymentDetails->getOperations() as $paymentDetail) {
                 if (!in_array($paymentDetail->getStatus(), array(
+                    self::STATUS_PAYMENT_CREATED,
                     self::STATUS_PAYMENT_REFUNDED,
                     self::STATUS_REFUND_REQUESTED,
                     self::STATUS_PAYMENT_REJECTED