+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.28] - 2026-15-04
+### Fixed
+- Remove CREATED operations from the operations display on the order details page
+### Changed
+- Correct display of payment method name for split payment transactions
+### Added
+- Add a retry mechanism for updating order status to Paid for captured webhooks
+
+## [1.4.27] - 2026-06-04
+### Changed
+- Separate shipping costs from product line items in the API request
+### Added
+- Added SECURITY.md file
+
+## [1.4.26] - 2026-31-03
+### Added
+- Added: Support of Illicado payment method
+
 ## [1.4.25] - 2026-26-03
 ### Fixed
 - Order incorrectly set to "Payment error" when delayed payment confirmation webhook arrives for an already created order